Move folding protocol type to Workspaces
